+++ examples/loco_example/src/controllers/notes.rs
@@ -63,7 +63,7 @@ pub fn routes() -> Routes {
         .prefix("notes")
         .add("/", get(list))
         .add("/", post(add))
-        .add("/:id", get(get_one))
-        .add("/:id", delete(remove))
-        .add("/:id", post(update))
+        .add("/{id}", get(get_one))
+        .add("/{id}", delete(remove))
+        .add("/{id}", post(update))
 }