+++ src/executor/query.rs
@@ -153,6 +153,56 @@ impl QueryResult {
             _ => unreachable!(),
         }
     }
+
+    /// Access the underlying `MySqlRow` if we use the MySQL backend.
+    #[cfg(feature = "sqlx-mysql")]
+    pub fn try_as_mysql_row(&self) -> Option<&sqlx::mysql::MySqlRow> {
+        match &self.row {
+            QueryResultRow::SqlxMySql(mysql_row) => Some(mysql_row),
+            #[allow(unreachable_patterns)]
+            _ => None,
+        }
+    }
+
+    /// Access the underlying `PgRow` if we use the Postgres backend.
+    #[cfg(feature = "sqlx-postgres")]
+    pub fn try_as_pg_row(&self) -> Option<&sqlx::postgres::PgRow> {
+        match &self.row {
+            QueryResultRow::SqlxPostgres(pg_row) => Some(pg_row),
+            #[allow(unreachable_patterns)]
+            _ => None,
+        }
+    }
+
+    /// Access the underlying `SqliteRow` if we use the SQLite backend.
+    #[cfg(feature = "sqlx-sqlite")]
+    pub fn try_as_sqlite_row(&self) -> Option<&sqlx::sqlite::SqliteRow> {
+        match &self.row {
+            QueryResultRow::SqlxSqlite(sqlite_row) => Some(sqlite_row),
+            #[allow(unreachable_patterns)]
+            _ => None,
+        }
+    }
+
+    /// Access the underlying `MockRow` if we use a mock.
+    #[cfg(feature = "mock")]
+    pub fn try_as_mock_row(&self) -> Option<&crate::MockRow> {
+        match &self.row {
+            QueryResultRow::Mock(mock_row) => Some(mock_row),
+            #[allow(unreachable_patterns)]
+            _ => None,
+        }
+    }
+
+    /// Access the underlying `ProxyRow` if we use a proxy.
+    #[cfg(feature = "proxy")]
+    pub fn try_as_proxy_row(&self) -> Option<&crate::ProxyRow> {
+        match &self.row {
+            QueryResultRow::Proxy(proxy_row) => Some(proxy_row),
+            #[allow(unreachable_patterns)]
+            _ => None,
+        }
+    }
 }
 
 #[allow(unused_variables)]
@@ -951,6 +1001,40 @@ mod postgres_array {
                 #[allow(unreachable_patterns)]
                 _ => unreachable!(),
             }
+        }
+    }
+}
+
+#[cfg(feature = "postgres-vector")]
+impl TryGetable for pgvector::Vector {
+    #[allow(unused_variables)]
+    fn try_get_by<I: ColIdx>(res: &QueryResult, idx: I) -> Result<Self, TryGetError> {
+        match &res.row {
+            #[cfg(feature = "sqlx-mysql")]
+            QueryResultRow::SqlxMySql(_) => {
+                Err(type_err("Vector unsupported by sqlx-mysql").into())
+            }
+            #[cfg(feature = "sqlx-postgres")]
+            QueryResultRow::SqlxPostgres(row) => row
+                .try_get::<Option<pgvector::Vector>, _>(idx.as_sqlx_postgres_index())
+                .map_err(|e| sqlx_error_to_query_err(e).into())
+                .and_then(|opt| opt.ok_or_else(|| err_null_idx_col(idx))),
+            #[cfg(feature = "sqlx-sqlite")]
+            QueryResultRow::SqlxSqlite(_) => {
+                Err(type_err("Vector unsupported by sqlx-sqlite").into())
+            }
+            #[cfg(feature = "mock")]
+            QueryResultRow::Mock(row) => row.try_get::<pgvector::Vector, _>(idx).map_err(|e| {
+                debug_print!("{:#?}", e.to_string());
+                err_null_idx_col(idx)
+            }),
+            #[cfg(feature = "proxy")]
+            QueryResultRow::Proxy(row) => row.try_get::<pgvector::Vector, _>(idx).map_err(|e| {
+                debug_print!("{:#?}", e.to_string());
+                err_null_idx_col(idx)
+            }),
+            #[allow(unreachable_patterns)]
+            _ => unreachable!(),
         }
     }
 }